24/06/2019

+++ UF2216/src/com/ipartek/formacion/uf2216/ejercicios/global/accesodatos/Crudable.java
@@ -12,7 +12,7 @@ public interface Crudable<T> {
 	public void insertar(T objeto);
 	public void modificar(T objeto);
 	public void borrar(T objeto);
-	public void borrar(long id);
+	public void borrar(int id);
 	public int numeroDeLibros();
 	
 }
